+++ src/video_core/renderer_opengl/gl_shader_manager.cpp
@@ -93,8 +93,9 @@ static void SetShaderUniformBlockBinding(GLuint shader, const char* name, Unifor
     }
     GLint ub_size = 0;
     glGetActiveUniformBlockiv(shader, ub_index, GL_UNIFORM_BLOCK_DATA_SIZE, &ub_size);
-    ASSERT_MSG(ub_size == expected_size, "Uniform block size did not match! Got {}, expected {}",
-               static_cast<int>(ub_size), expected_size);
+    ASSERT_MSG(static_cast<std::size_t>(ub_size) == expected_size,
+               "Uniform block size did not match! Got {}, expected {}", static_cast<int>(ub_size),
+               expected_size);
     glUniformBlockBinding(shader, ub_index, static_cast<GLuint>(binding));
 }
